requirements.txt
aiohttp>=3.9.0,<4.0.0
msgspec>=0.19.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
curl_cffi>=0.7.0,<1.0.0
bilibili-api-python>=17.4.1,<18.0.0
